requirements.txt
Django==1.11.15
django-cors-headers==2.4.0
django-storages==1.6.6
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
mysqlclient==1.3.13
